🐛 Bugs fixes

+++ commands/Config/membercount.js
@@ -3,47 +3,47 @@ module.exports.run = async (client, message, args, data) => {
         if(data.plugins.membercount?.parentID && message.guild.channels.resolve(data.plugins.membercount?.parentID)) {
             return message.channel.send(`⚠️ Les salons compteur de membre existent déjà ! Faites \`${data.prefix}membercount delete\` pour le retirer`);
         } else {
-            try {
-                if(message.guild.channels.cache.size >= 496) return message.channel.send('⚠️ Vous avez atteint la limite de salon (500). Je ne peux pas en créer tant qu\'il y en a plus que 496.');
-                if(message.guild.memberCount !== message.guild.members.cache.size) await message.guild.members.fetch();
+            if(message.guild.channels.cache.size > 495) return message.channel.send('⚠️ Vous avez atteint la limite de salon (500). Je ne peux pas en créer tant qu\'il y en a plus que 495.');
+            if(message.guild.memberCount !== message.guild.members.cache.size) await message.guild.members.fetch();
 
-                const prompts = [
-                    'Quel nom souhaitez vous donner aux salons de membres (humains) ?\nExemple: `👨 Humains: {count}`',
-                    'Quel nom souhaitez vous donner aux salons de membres (bots) ?\nExemple: `🤖 Bots: {count}`',
-                    'Quel nom souhaitez vous donner aux salons de membres (total) ?\nExemple: `🌎 Total: {count}`'
-                ];
-                const filter = (m) => m.author.id === message.author.id;
+            const prompts = [
+                'Quel nom souhaitez vous donner aux salons de membres (humains) ?\nExemple: `👨 Humains: {count}`',
+                'Quel nom souhaitez vous donner aux salons de membres (bots) ?\nExemple: `🤖 Bots: {count}`',
+                'Quel nom souhaitez vous donner aux salons de membres (total) ?\nExemple: `🌎 Total: {count}`'
+            ];
+            const filter = (m) => m.author.id === message.author.id;
 
-                const membersChannelText = await message.channel.send(prompts[0]);
+            const membersChannelText = await message.channel.send(prompts[0]);
 
-                message.channel.awaitMessages(filter, { time: 30000, max: 1 })
-                    .then(async (membersChannelName) => {
-                        if(!membersChannelName.first().content) return message.channel.send('⚠️ Merci de spécifier un nom.');
-                        if(!membersChannelName.first().content.includes('{count}')) return message.channel.send('⚠️ Vous devez mettre la balise de nombre de membre `{count}`. Sans elle, les salons compteurs ne serviraient à rien...');
+            message.channel.awaitMessages(filter, { time: 30000, max: 1 })
+                .then(async (membersChannelName) => {
+                    if(!membersChannelName.first().content) return message.channel.send('⚠️ Merci de spécifier un nom.');
+                    if(!membersChannelName.first().content.includes('{count}')) return message.channel.send('⚠️ Vous devez mettre la balise de nombre de membre `{count}`. Sans elle, les salons compteurs ne serviraient à rien...');
 
-                        membersChannelName.first().delete().catch(() => {});
-                        membersChannelText.delete().catch(() => {});
+                    membersChannelName.first().delete().catch(() => {});
+                    membersChannelText.delete().catch(() => {});
 
-                        const botsChannelText = await message.channel.send(prompts[1]);
+                    const botsChannelText = await message.channel.send(prompts[1]);
 
-                        message.channel.awaitMessages(filter, { time: 30000, max: 1 })
-                            .then(async (botsChannelName) => {
-                                if(!botsChannelName.first().content) return message.channel.send('⚠️ Merci de spécifier un nom.');
-                                if(!botsChannelName.first().content.includes('{count}')) return message.channel.send('⚠️ Vous devez mettre la balise de nombre de membre `{count}`. Sans elle, les salons compteurs ne serviraient à rien...');
-    
-                                botsChannelName.first().delete().catch(() => {});
-                                botsChannelText.delete().catch(() => {});
+                    message.channel.awaitMessages(filter, { time: 30000, max: 1 })
+                        .then(async (botsChannelName) => {
+                            if(!botsChannelName.first().content) return message.channel.send('⚠️ Merci de spécifier un nom.');
+                            if(!botsChannelName.first().content.includes('{count}')) return message.channel.send('⚠️ Vous devez mettre la balise de nombre de membre `{count}`. Sans elle, les salons compteurs ne serviraient à rien...');
 
-                                const totalChannelText = await message.channel.send(prompts[2]);
+                            botsChannelName.first().delete().catch(() => {});
+                            botsChannelText.delete().catch(() => {});
 
-                                message.channel.awaitMessages(filter, { time: 30000, max: 1 })
-                                    .then(async (totalChannelName) => {
-                                        if(!totalChannelName.first().content) return message.channel.send('⚠️ Merci de spécifier un nom.');
-                                        if(!totalChannelName.first().content.includes('{count}')) return message.channel.send('⚠️ Vous devez mettre la balise de nombre de membre `{count}`. Sans elle, les salons compteurs ne serviraient à rien...');
-            
-                                        totalChannelName.first().delete().catch(() => {});
-                                        totalChannelText.delete().catch(() => {});
+                            const totalChannelText = await message.channel.send(prompts[2]);
 
+                            message.channel.awaitMessages(filter, { time: 30000, max: 1 })
+                                .then(async (totalChannelName) => {
+                                    if(!totalChannelName.first().content) return message.channel.send('⚠️ Merci de spécifier un nom.');
+                                    if(!totalChannelName.first().content.includes('{count}')) return message.channel.send('⚠️ Vous devez mettre la balise de nombre de membre `{count}`. Sans elle, les salons compteurs ne serviraient à rien...');
+
+                                    totalChannelName.first().delete().catch(() => {});
+                                    totalChannelText.delete().catch(() => {});
+
+                                    try {
                                         const parent = await message.guild.channels.create('Stats du serveur', {
                                             type: 'category',
                                             permissionOverwrites: [
@@ -51,6 +51,10 @@ module.exports.run = async (client, message, args, data) => {
                                                     id: message.guild.roles.everyone.id,
                                                     allow: ['VIEW_CHANNEL'],
                                                     deny: ['CONNECT']
+                                                },
+                                                {
+                                                    id: client.user.id,
+                                                    allow: ['VIEW_CHANNEL', 'MANAGE_CHANNELS']
                                                 }
                                             ],
                                             position: 0
@@ -64,7 +68,6 @@ module.exports.run = async (client, message, args, data) => {
                                         const botsChannel = await message.guild.channels.create(botsChannelName.first().content.replace('{count}', message.guild.members.cache.filter(m => m.user.bot).size), {
                                             type: 'voice',
                                             parent
-
                                         });
 
                                         const totalChannel = await message.guild.channels.create(totalChannelName.first().content.replace('{count}', message.guild.memberCount), {
@@ -94,47 +97,59 @@ module.exports.run = async (client, message, args, data) => {
                                         await data.save();
 
                                         message.channel.send('✅ Les salons de compteur de membre ont bien été mis en place.');
-                                    })
-                                    .catch(() => message.channel.send('Temps écoulé'));
-                            })
-                            .catch(() => message.channel.send('Temps écoulé'));
-                    })
-                    .catch(() => message.channel.send('Temps écoulé'));
-            } catch (e) {
-                console.error(e);
-                client.channels.cache.get(client.config.support.logs).send(`Une erreur est survenue lors de la commande membercount: \n\`\`\`${e.stack}\`\`\``);
-                return message.channel.send(`Une erreur est survenue lors de la création des salons: \n\`\`\`${e.message}\`\`\``);
-            }
+                                    } catch (e) {
+                                        console.error(e);
+                                        client.channels.cache.get(client.config.support.logs).send(`Une erreur est survenue lors de la commande membercount: \n\`\`\`${e.stack}\`\`\``);
+                                        return message.channel.send(`Une erreur est survenue lors de la création des salons: \n\`\`\`${e.message}\`\`\``);
+                                    }
+                                })
+                                .catch(() => message.channel.send('Temps écoulé'));
+                        })
+                        .catch(() => message.channel.send('Temps écoulé'));
+                })
+                .catch(() => message.channel.send('Temps écoulé'));
         }
     } else if(args[0].toLowerCase() === 'delete') {
         if(!data.plugins.membercount?.parentID || !message.guild.channels.resolve(data.plugins.membercount?.parentID)) return message.channel.send(`⚠️ Les salons de compteur de membres ne sont pas activés sur ce serveur, faites \`${data.prefix}membercount create\` pour les créer !`);
 
-        const deleteChannel = (channel) => setTimeout(async () => await message.guild.channels.resolve(channel)?.delete().catch(() => {}), 1000);
+        const channels = Object.values(data.plugins.membercount.channels);
+
+        if(channels.filter((ch) => message.guild.channels.cache.get(ch.id)).some((channel) => !message.guild.me.permissionsIn(channel.id).has(['VIEW_CHANNEL', 'MANAGE_CHANNELS']))) return message.channel.send('⚠️ Je n\'ai pas les permissions de supprimer l\'un des salons vocaux de compteur de membres.');
+        const deleteChannel = (channel) => setTimeout(async () => await message.guild.channels.cache.get(channel).delete().catch(() => {}), 1000);
 
         deleteChannel(data.plugins.membercount.parentID);
-        Object.values(data.plugins.membercount.channels).forEach((channel) => deleteChannel(channel.id));
+        channels.forEach((channel) => deleteChannel(channel.id));
 
-        data.plugins.membercount = {
-            channels: {
-                members: {
-                    name: null,
-                    id: null
+        setTimeout(async () => {
+            let notDeleted = 0;
+            Object.values(data.plugins.membercount.channels).forEach((chnl) => {
+                const ch = message.guild.channels.cache.get(chnl.id);
+                if(ch && !ch.deleted) notDeleted++;
+            });
+    
+            data.plugins.membercount = {
+                channels: {
+                    members: {
+                        name: null,
+                        id: null
+                    },
+                    bots: {
+                        name: null,
+                        id: null
+                    },
+                    totalMembers: {
+                        name: null,
+                        id: null
+                    }
                 },
-                bots: {
-                    name: null,
-                    id: null
-                },
-                totalMembers: {
-                    name: null,
-                    id: null
-                }
-            },
-            parentID: null
-        };
-        data.markModified("plugins.membercount");
-
-        await data.save();
-        setTimeout(() => message.channel.send('✅ Les salons de compteur de membres ont été supprimés'), 1500);
+                parentID: null
+            };
+            data.markModified("plugins.membercount");
+    
+            await data.save();
+            if(notDeleted >= 1) return message.channel.send(`⚠️ Je n'ai pas réussi à supprimer ${notDeleted} salon sur 4.`);
+            else message.channel.send('✅ Les salons de compteur de membres ont été supprimés.');
+        }, 1500);
     } else message.channel.send(`⚠️ Vous n'utilisez pas la commande correctement.\nFaites \`${data.prefix}membercount create\` pour créer le compteur de membre, et \`${data.prefix}membercount delete\` pour le supprimer.`);
 }
